--- libs/optimizer/src/index.ts
@@ -1,3 +1,0 @@
-export { getOptimizer, CircuitType } from './lib/optimizer';
-export type { LatLonAltTime, ScoringTrack, ScoringResult, ScoringRequest } from './lib/optimizer';
-export type { ScoringRuleName, scoringRuleNames } from './lib/scoring-rules';
