+++ libs/optimizer/src/lib/scoring-rules.ts
@@ -1,24 +1,8 @@
 import * as igcXcScore from 'igc-xc-score';
-
-export const scoringRuleNames = [
-  'CzechLocal',
-  'CzechEurope',
-  'CzechOutsideEurope',
-  'FFVL',
-  'Leonardo',
-  'Norway',
-  'UKClub',
-  'UKInternational',
-  'UKNational',
-  'XContest',
-  'XContestPPG',
-  'WorldXC',
-] as const;
-
-export type ScoringRuleName = (typeof scoringRuleNames)[number];
+import type { ScoringRuleName } from './api';
 
 // TODO: Export the rules from igc-xc-score
-const scoringBaseModel = igcXcScore.scoringRules['XContest'];
+const scoringBaseModel = igcXcScore.scoringRules.XContest;
 const openDistance = scoringBaseModel[0];
 const freeTriangle = scoringBaseModel[1];
 const faiTriangle = scoringBaseModel[2];
